+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
-- Added `margin` and `padding` shorthands
+- Added `margin` and `padding` shorthands.
+- Added `borderWidth` shorthand.
 
 ## v0.1.0 (August 19, 2023)